+++ src/pages/Public/News/styles/News.component.scss
